pyproject.toml
[tool.poetry]
name = "template-matching"
version = "0.1.0"
description = ""
authors = ["hiro16131104 <hiro931104@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
black = "^24.4.2"
flake8 = "^7.1.0"
isort = "^5.13.2"
opencv-python = "^4.10.0.84"
numpy = "^2.0.1"
matplotlib = "^3.9.1"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"